requirements.txt
insightface==0.7.3
opencv-python-headless==4.9.0.80
pillow>=10.3.0
scikit-learn>=1.4.0
umap-learn>=0.5.6
chinese-whispers>=0.8.0
networkx>=3.3
PyQt6>=6.7.0
numpy>=1.26.0
Pillow-HEIF>=0.16.0
